some small changes
